removed code smell, moved remediation mock to shared folder
